feat(FN-NONE): move incorrectly defined api tests
