+++ dtfs-central-api/src/utils/get-object-property-value-from-string-path.test.js
@@ -1,4 +1,4 @@
-const getObjectPropertyValueFromStringPath = require('../../../src/utils/getObjectPropertyValueFromStringPath');
+const getObjectPropertyValueFromStringPath = require('./getObjectPropertyValueFromStringPath');
 
 describe('getObjectPropertyValueFromStringPath', () => {
   it('should return a property in object from the given string path', () => {
